+++ Assets/GameMain/DataTables/SoDataTable/Scripts/WeaponDataTable.cs
@@ -11,17 +11,17 @@ public class WeaponDataRow:SoDataRow
     public float attackInterval;
     [Header("属性加减,Add为数值增加，Times为乘算")]
     public int hpAdd;
-    public float hpTimes;
+    public int hpTimes;
     public int mpAdd;
-    public float mpTimes;
+    public int mpTimes;
     public int attackAdd;
-    public float attackTimes;
+    public int attackTimes;
     public int defenseAdd;
-    public float defenseTimes;
+    public int defenseTimes;
     public int moveSpeedAdd;
-    public float moveSpeedTimes;
+    public int moveSpeedTimes;
     public int actionSpeedAdd;
-    public float actionSpeedTimes;
+    public int actionSpeedTimes;
     [Header("填写武器自带技能和buff，\n此项目左键放skills[0]，右键skills[1]")]
     public string[] skills;
     public string[] buffs;
